requirements.txt
redis==2.10.5
elasticsearch==2.4.0
requests==2.7.0
beautifulsoup4==4.5.1
PyYAML==3.12